Commands/b_star/requirements.txt
discord.py==1.7.3
numpy==1.22.0
Pillow==9.0.1
psycopg2==2.9.2
requests==2.23.0
python-dotenv==0.19.1
boto3==1.24.12
func_timeout==4.3.5
lark==1.1.2